fix: Permitir cantidades decimales en transferencias de almacén
- Agregar botón de punto decimal al numpad
- Cambiar parseInt a parseFloat para soportar decimales
- Estilizar botón decimal con color azul

+++ src/app/dashboard/market/warehouses/[id]/operations/page.tsx
@@ -715,7 +715,7 @@ export default function WarehouseOperationsPage() {
 
   const handleNumpadConfirm = () => {
     if (selectedProductIndex !== null && operation.products[selectedProductIndex]) {
-      const quantity = parseInt(numpadValue) || 1
+      const quantity = parseFloat(numpadValue) || 1
       const productId = operation.products[selectedProductIndex].id
 
       if (operation.operationType === 'adjustment') {
